@@ -1,6 +1,6 @@
 import { createElement } from 'react';
-import ReactDOMFactories from '../node_modules/react/lib/ReactDOMFactories';
-import { isFunction, isArray, forEach, omit, clone, has, isObject, isString } from 'lodash';
+import ReactDOMFactories from 'react/lib/ReactDOMFactories';
+import { isArray, forEach, omit, clone, has, isObject } from 'lodash';
 
 let _componentMap = null;
 
@@ -44,7 +44,7 @@ export default class ReactJsonSchema {
       } else if (has(ReactDOMFactories, schema.component)) {
         Component = schema.component;
       }
-    }  else {
+    } else {
       throw new Error('ReactJsonSchema could not resolve a component due to a missing component attribute in the schema.');
     }
     return Component;
